Rock Paper Scissors Lizard Spock feedback
