+++ utils/test/feedback/rock_paper_scissors_lizard_spock_test.exs
@@ -1,0 +1,5 @@
+defmodule Utils.Feedback.RockPaperScissorsLizardSpockTest do
+  use Utils.FeedbackCase,
+    module: Utils.Feedback.RockPaperScissorsLizardSpock,
+    file: "../exercises/rock_paper_scissors_lizard_spock.livemd"
+end
